untitled5/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
qrcode[pil]
pillow
werkzeug